rename life cycle phase components
