+++ app/components/projects/settings/life_cycle/index_page_header_component.html.erb
