+++ app/components/projects/settings/life_cycle/phase_component.html.erb
