+++ app/views/projects/settings/life_cycle_steps/index.html.erb
@@ -27,10 +27,10 @@ See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
 
-<%= render(Projects::Settings::LifeCycleSteps::IndexPageHeaderComponent.new(project: @project)) %>
+<%= render(Projects::Settings::LifeCycle::IndexPageHeaderComponent.new(project: @project)) %>
 
 <%= render(
-      Projects::Settings::LifeCycleSteps::IndexComponent.new(
+      Projects::Settings::LifeCycle::IndexComponent.new(
         project: @project,
         life_cycle_definitions: @life_cycle_definitions
       )
